requirements.txt
# Core dependencies (2026년 최신 버전)
openai>=1.65.0  # GPT-5.2, Structured Outputs, 1M context 지원

# Optional: Anthropic Claude support
# anthropic>=0.45.0  # Claude Opus 4.5, Sonnet 4.5 지원

# Database drivers
# psycopg2-binary>=2.9.9  # PostgreSQL
# google-cloud-bigquery>=3.15.0  # BigQuery
# snowflake-connector-python>=3.8.0  # Snowflake

# Embedding 기반 스키마 링킹 (선택적)
# numpy>=1.26.0
# tiktoken>=0.6.0

# Utilities
python-dotenv>=1.0.1

# Development (optional)
# pytest>=8.0.0
# black>=24.0.0
# mypy>=1.8.0
# ruff>=0.2.0  # 최신 린터
